refactor(agent): split template rendering from commands
